requirements_01.txt
da-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/da_core_news_sm-2.3.0/da_core_news_sm-2.3.0.tar.gz
# install manually with: python -m spacy download da_core_news_sm
aiofiles==0.8.0
datasets==1.17.0
fastapi==0.68.1
joblib==1.0.1
laserembeddings==1.1.1
loguru==0.5.3
matplotlib==3.4.3
numpy==1.19.5
pandas==1.3.2
project_template==0.1.0
pyarrow==6.0.1
pytest==6.2.5
python-dotenv==0.19.2
python_dateutil==2.8.2
scikit_learn==1.0.1
seaborn==0.11.2
sentence_transformers==2.1.0
setuptools==59.4.0
spacy==2.3.5
stop_words==2018.7.23
keras==2.6.*
tensorflow-cpu==2.6.0
transformers==4.11.3
uvicorn==0.15.0
xgboost==1.4.2
